update build to deal with warnings

+++ build.ps1
@@ -1,15 +1,21 @@
-rm -Force -Recurse artifacts
+if (Test-Path artifacts) {
+	rm -Force -Recurse artifacts
+}
 
 dotnet restore .\templates
 dotnet pack .\templates\templates.csproj -o .\artifacts\ --no-build
 
-rm -Force -Recurse temp
-dotnet new --uninstall cortside.templates
-dotnet new --install .\artifacts\*.nupkg
+if (Test-Path temp) {
+	rm -Force -Recurse temp
+}
+dotnet new uninstall cortside.templates
+dotnet new install .\artifacts\*.nupkg
 
 dotnet new cortside-api --output ./temp --name Foo.Bar --company Foo --product Bar
 
 dotnet build ./temp/src
 dotnet test ./temp/src
 
-rm -Force -Recurse temp
+if (Test-Path temp) {
+	rm -Force -Recurse temp
+}
